rename authRequest to Authorization model
