+++ api/app/Actions/Lecture/LectureCollectionAction.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Actions\Lecture;
 
-use App\Actions\Auth\AuthRequest;
+use App\Models\Authentication;
 use App\Repository\{LectureRepository, UserRepository};
 
 final class LectureCollectionAction
@@ -15,7 +15,7 @@ final class LectureCollectionAction
 
     public function execute(LectureCollectionRequest $request): LectureCollectionResponse
     {
-        $response = AuthRequest::isUserLecturer() ?
+        $response = Authentication::isUserLecturer() ?
             $this->lectureRepository->getLecturesByLecturer($request->getUserId()) :
             $this->userRepository->getLecturesByUser($request->getUserId());
